fix: critical bug fixes from full audit of solidus-core
- PayCommand: replace manual deduct+add with atomic transferOffline()
  for offline payments, eliminating negative balance exploit and
  fire-and-forget refund bug (money loss)
- PayCommand: add null-check for server in async callback (stale ref)
- SellCommand: skip armor slots (36-39) in /sell all to prevent
  accidental armor sales that contradict documented behavior
- SellCommand: remove overly broad reverse-direction partial match
  in item name search (e.g., searching 'D' matching everything)
- ShopGUI: fix duplicate section icons when >9 sections — was using
  slotIndex instead of entryIndex to skip already-placed sections
- TransactionsCommand: hide next-page hint on last page
- SolidusMod: change tickCounter from int to long to prevent
  integer overflow on long-running servers

+++ src/main/java/com/solidus/SolidusMod.java
@@ -51,7 +51,7 @@ public class SolidusMod implements DedicatedServerModInitializer {
     private static RateLimiter rateLimiter;
 
     /** Tick counter for periodic tasks (auction expiration check every 5 minutes) */
-    private static int tickCounter = 0;
+    private static long tickCounter = 0;
     private static final int AUCTION_EXPIRY_CHECK_INTERVAL = 6000; // 5 minutes (6000 ticks)
 
     @Override
